move custom elements fetcher to helper
