+++ packages/components/.storybook/preview.js
@@ -1,45 +1,8 @@
-import { setCustomElementsManifest } from '@storybook/web-components';
 import 'normalize.css';
 import '../src/solid-styles.css';
 import '../src/styles/tailwind.css';
-import { fetchStyleComponents } from '../scripts/storybook/styles-helper';
 import { registerIconLibrary } from '../src/utilities/icon-library';
 import { storybookUtilities } from '../scripts/storybook/helper';
-
-/**
- * This loads the custom elements manifest generated on run or on build time.
- */
-
-async function loadCustomElements() {
-  let customElements;
-
-  fetch('./custom-elements.json')
-    .then(() => {
-      // Use dynamic manifest generated by Vite on runtime
-      return fetch('./custom-elements.json').then(res => res.json());
-    })
-    .catch(() => {
-      console.log('Failed to fetch custom-elements.json. Using local manifest...');
-      // Use manifest file generated on build time
-      return import('../dist/custom-elements.json');
-    })
-    .then(async customElements => {
-      // In production, the manifest is optimized and styles are added during the build process.
-      // However, in development mode, the manifest isn't built with styles, so we add them dynamically.
-      // In the latter case therefore we have to "fetchStyleComponents", too.
-      const stylesAreAlreadyIncluded = customElements.modules.some(module => module.styles === true);
-      if (!stylesAreAlreadyIncluded) {
-        const styleComponents = await fetchStyleComponents();
-        customElements.modules = [...customElements?.modules, ...styleComponents];
-      }
-
-      setCustomElementsManifest(customElements);
-    });
-
-  console.log('Custom elements manifest loaded');
-}
-
-await loadCustomElements();
 
 /**
  * This registers iconLibraries for the sd-icon component
