+++ packages/components/scripts/storybook/fetch-cem.ts
@@ -1,0 +1,30 @@
+import { fetchStyleComponents } from './styles-helper';
+import { setCustomElementsManifest } from '@storybook/web-components';
+
+async function fetchCustomElementsManifest() {
+  try {
+    await fetch('./custom-elements.json');
+
+    // Use dynamic manifest generated by Vite on runtime
+    const response = await fetch('./custom-elements.json');
+    return await response.json();
+  } catch (error) {
+    console.log('Failed to fetch custom-elements.json. Using local manifest...');
+
+    // Use manifest file generated on build time
+    return await import('../../dist/custom-elements.json');
+  }
+}
+
+export default async function loadCustomElements() {
+  const customElements = await fetchCustomElementsManifest();
+
+  const stylesAreAlreadyIncluded = customElements.modules.some(module => module.styles === true);
+  if (!stylesAreAlreadyIncluded) {
+    const styleComponents = await fetchStyleComponents();
+    customElements.modules = [...customElements?.modules, ...styleComponents];
+  }
+
+  setCustomElementsManifest(customElements);
+  console.log('Custom elements manifest loaded');
+}
